+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## [0.0.11] Push Notification
 * handleReceivedMessage method added to get chat message from FCM Notification
-* authToken Deprecated insteadof use refreshAndGetAuthToken
+* authToken Deprecated instead of use refreshAndGetAuthToken
 
 ## [0.0.10] Code Optimization
 * Minified Enabled True Support Added